+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-acdc/index.js
@@ -1,0 +1,33 @@
+import template from './swag-paypal-acdc.html.twig';
+
+const { Component } = Shopware;
+
+Component.register('swag-paypal-acdc', {
+    template,
+
+    inject: [
+        'acl',
+    ],
+
+    props: {
+        actualConfigData: {
+            type: Object,
+            required: true,
+        },
+        allConfigs: {
+            type: Object,
+            required: true,
+        },
+        selectedSalesChannelId: {
+            type: String,
+            required: false,
+            default: null,
+        },
+    },
+
+    methods: {
+        checkBoolFieldInheritance(value) {
+            return typeof value !== 'boolean';
+        },
+    },
+});